+++ AeSys/EoDlgViewParameters.h
@@ -22,7 +22,7 @@ class EoDlgViewParameters : public CDialog {
 
  public:
   BOOL m_PerspectiveProjection;
-  unsigned long m_ModelView;
+  uintptr_t m_ModelView;
   afx_msg void OnBnClickedApply();
   afx_msg void OnEnChangePositionX();
   afx_msg void OnEnChangePositionY();
